:art: style: 코드 수정 - comments serializer simpleuser 삭제 - search filters.py와 분리 - search 독스트링 작성
